Datos sobre pacientes diabeticos

Archivo de datos con informacion sobre algunas variables antropometricas denigraficas y clinicas de un conjunto de 94 diabeticos (por compatibilidad se omiten acentos y caracteres no anglosajones)

Usage

osteo

Format

data frame con 27 variables y 94 casos:

num

numero de historia,

edad

edad del paciente, en anos

grupo_edad

categoria de edad del paciente, en anos

sexo

sexo del paciente

peso

peso del paciente, en kg

talla

talla del paciente, en cm

imc

indice de masa corporal de Quetelet

tevol

tiempo de evolucion de la enfermedad, en anos

tabaco

variable binaria, indicadora de si el paciente es fumador

alcohol

variable binaria, indicadora de si el paciente bebe alcohol o no

ingca

variable binaria, indicadora de si el paciente toma suficiente calcio o no

acfis

variable binaria, indicadora de la realizacion de actividad fisica regular

retin

presencia y gravedad de retinopatia

nefro

presencia y gravedad de nefropatia

neuro

presencia y gravedad de neuropatia

hba1c

porcentaje de hemoglobina glicosilada

ca

nivel de calcio, mg/dL

p

nivel de fosforo, mg/dL

cr

nivel de creatinina, mg/dL

pthm

nivel de hormona paratiroidea media, ng/ml

pthi

nivel de hormona paratiroidea intacta, pg/ml

bmdcue

densidad de masa osea en el cuello del femur

szl24

densidad de masa osea tipificada en l24

szl24

densidad de masa osea tipificada en el triangulo de Ward

szcue

densidad de masa osea tipificada en el cuello del femur

osteo_cue

presencia de osteoporosis en el cuello del femur

osteo_tri

presencia de osteoporosis en el triangulo

Source

http://www.ugr.es/local/bioest/osteo.RData